requirements.txt
asgiref==3.2.3
beautifulsoup4==4.9.0
bootstrap4==0.1.0
certifi==2019.11.28
cffi==1.14.0
configparser==5.0.0
Django==3.0.3
django-bootstrap4==1.1.1
django-braces==1.14.0
django-debug-toolbar==2.2
gunicorn==20.0.4
misaka==2.1.1
mysql-connector==2.2.9
pycparser==2.20
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
